Refactored Kafka config a bit
