+++ config/dev/cards-publication-dev.yml
@@ -3,28 +3,22 @@ server:
 spring:
   application:
     name: cards-publication
-  deserializer:
-    value:
-      delegate:
-        class: io.confluent.kafka.serializers.KafkaAvroDeserializer
-  serializer:
-    value:
-      delegate:
-        class: io.confluent.kafka.serializers.KafkaAvroSerializer
 
-  # uncomment to enable Kafka
-#  kafka:
-#    consumer:
+  # uncomment kafka.consumer.group-id to enable Kafka
+  kafka:
+    consumer:
 #      group-id: opfab-command
-
-schema:
-  registry:
-    url: http://localhost:8081
+      value-deserializer: io.confluent.kafka.serializers.KafkaAvroDeserializer
+    producer:
+      value-serializer: io.confluent.kafka.serializers.KafkaAvroSerializer
 
 opfab:
   kafka:
     topics:
       topicname: opfab
+    schema:
+      registry:
+        url: http://localhost:8081
 
 #here we put urls for all feign clients
 users:
