+++ services/core/cards-publication/src/main/java/org/lfenergy/operatorfabric/cards/publication/configuration/kafka/ConsumerFactoryAutoConfiguration.java
@@ -14,7 +14,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.kafka.clients.consumer.ConsumerConfig;
 import org.lfenergy.operatorfabric.avro.CardCommand;
-import org.lfenergy.operatorfabric.cards.publication.kafka.consumer.SchemaRegistryProperties;
+import org.lfenergy.operatorfabric.cards.publication.kafka.SchemaRegistryProperties;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.boot.autoconfigure.kafka.KafkaProperties;
@@ -40,8 +40,8 @@ public class ConsumerFactoryAutoConfiguration {
 
     private String deserializerKeyClass = "org.apache.kafka.common.serialization.StringDeserializer";
 
-    @Value("${spring.deserializer.value.delegate.class}")
-    private String deserializerValueClass;
+    @Value("${spring.kafka.consumer.value-deserializer}")
+    private String valueDeserializer;
 
     private Map<String,Object> consumerConfig() {
         log.info("bootstrapServers: " + kafkaProperties.getBootstrapServers());
@@ -49,7 +49,7 @@ public class ConsumerFactoryAutoConfiguration {
         props.put(ConsumerConfig.KEY_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
         props.put(ErrorHandlingDeserializer.KEY_DESERIALIZER_CLASS, deserializerKeyClass);
         props.put(ConsumerConfig.VALUE_DESERIALIZER_CLASS_CONFIG, ErrorHandlingDeserializer.class);
-        props.put(ErrorHandlingDeserializer.VALUE_DESERIALIZER_CLASS, deserializerValueClass);
+        props.put(ErrorHandlingDeserializer.VALUE_DESERIALIZER_CLASS, valueDeserializer);
         props.put("specific.avro.reader", "true");
         return props;
     }
